+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Select: Search only the columns which can hold the searched value, convert the other types to text (bug #1216)
 - Select: Modify or delete several rows in a transaction
 - Edit: Keep the JSON value entered by the user after an error, it was replaced by null if it was invalid
+- Edit: Print the query fetching the row, the same as in Select
 - MariaDB 12: Support triggers for multiple events and for UPDATE OF (bug #1306)
 - PostgreSQL: Edit and delete rows identified by a column of a composite type (bug #1217)
 - MS SQL: Fix importing a CSV file to a table with an identity column through PDO, it silently did nothing
@@ -22,6 +23,8 @@
 ### Plugins
 - Method error() to modify the printed database error messages (bug #1249)
 - Method dumpPrint() to print HTML code in the export form
+- Method editRowPrint() receives the query which got the edited row and the time it took
+- Driver plugins: Driver::select() assigns the executed query to $this->query
 - Driver plugins: Bundle the syntax highlighter in the plugin file, adminer-plugins/jush-&lt;driver>.js is not loaded anymore
 - Driver plugins: Method jushAutocomplete() to disable or replace the autocomplete of the query &lt;textarea>
 - New plugin: Link the plugins usable in each part of the interface